+++ cmd.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 const resolve = require('resolve-from')
-const { dirname } = require('path')
+const { dirname, join } = require('path')
 const cwd = process.cwd()
 
 const optimize = loadOptimizer()
@@ -13,16 +13,16 @@ optimize.install({
 
 loadNextCmd()
 
-
 function loadNextCmd () {
   const pkg = resolve(cwd, 'next/package.json')
   const dir = dirname(pkg)
   const { bin } = require(pkg)
-  require(resolve(dir, bin.next))
+
+  require(join(dir, bin.next))
 }
 
 function loadOptimizer () {
-  try { 
+  try {
     return require(resolve(cwd, 'esx/optimize'))
   } catch (e) {
     return require(resolve(cwd, 'esx/experimental-optimize'))